new camera_keep_device_on() function (less noise in debug log)
Instead of calling ptp_canon_eos_keepdeviceon() in e.g.
camera_wait_for_event() directly for every invocation (often twice),
which can lead to dozens of those calls per second, this patch limits the
number to 1 every 10 seconds.

The camera_prepare_capture + ptp_check_eos_events + keepdeviceon block in
camera_lookup_by_property has been removed completely as this function is
only called from within camera_wait_for_event() where those checks have
been made just prior to the call to camera_lookup_by_property.

Furthermore: ptp_check_eos_events() will now only be called inside
camera_wait_for_event if the queue of already transmitted events is
drained first.
